requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
pydantic[email]
PyJWT
python-multipart